+++ AngularClient/src/dev-all/app.component.ts
@@ -1,7 +1,6 @@
 import { Component } from "@angular/core";
 
-import { User } from "forcrowd-backbone";
-import { AuthService } from "../main/core/core.module";
+import { User, AuthService } from "forcrowd-backbone";
 
 @Component({
   selector: "app",