+++ AngularClient/src/dev-all/app.module.ts
@@ -2,7 +2,8 @@ import { NgModule } from "@angular/core";
 
 import { AdminModule } from "../main/admin/admin.module";
 import { CoreModule } from "../main/core/core.module";
-import { SharedModule } from "forcrowd-backbone";
+import { SharedModule } from "../main/shared/shared.module";
+import { ForcrowdBackboneModule } from "forcrowd-backbone"
 
 import { AppRouterModule } from "./app-router.module";
 import { AppComponent } from "./app.component";
@@ -47,7 +48,8 @@ import { WebApiComponent } from "./web-api.component";
 
     CoreModule,
 
-    AdminModule
+    AdminModule,
+    ForcrowdBackboneModule
   ]
 })
 export class AppModule { }