+++ AngularClient/src/dev-all/odata.component.ts
@@ -1,7 +1,6 @@
 import { Component } from "@angular/core";
 
-import { AuthService } from "../main/core/core.module";
-import { User } from "forcrowd-backbone";
+import { AuthService, User } from "forcrowd-backbone";
 
 @Component({
   selector: "odata",