+++ AngularClient/src/main/account/account.service.ts
@@ -4,8 +4,7 @@ import { Observable } from "rxjs";
 import { map } from "rxjs/operators";
 
 import { AppSettings } from "../../app-settings/app-settings";
-import { User } from "forcrowd-backbone";
-import { AppHttpClient, AuthService } from "../core/core.module";
+import { AppHttpClient, AuthService, User } from "forcrowd-backbone";
 
 @Injectable()
 export class AccountService {