+++ AngularClient/src/main/account/change-password.component.ts
@@ -3,7 +3,7 @@ import { Router } from "@angular/router";
 
 import { AppSettings } from "../../app-settings/app-settings";
 import { AccountService } from "./account.service";
-import { NotificationService } from "../core/core.module";
+import { NotificationService } from "forcrowd-backbone";
 
 @Component({
   selector: "change-password",