+++ AngularClient/src/main/account/register.component.ts
@@ -3,8 +3,7 @@ import { Router } from "@angular/router";
 import { Subscription } from "rxjs";
 
 import { AppSettings } from "../../app-settings/app-settings";
-import { AuthService, NotificationService } from "../core/core.module";
-import { getUniqueEmail, getUniqueUserName, stripInvalidChars } from "forcrowd-backbone";
+import { AuthService, NotificationService, getUniqueEmail, getUniqueUserName, stripInvalidChars } from "forcrowd-backbone";
 
 @Component({
   selector: "register",